backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.3.2
pymongo==4.8.0
beanie==1.25.0
pydantic==2.10.3
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.1
shap==0.46.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
gunicorn==21.2.0
email-validator==2.1.0
psutil==6.1.1
